Create Update
This is an enhanced update to the index.js code I was working on before. I changed the movement so that the background won't move and the player will move around instead, plus an npc that has collision, can not overlap with the player, and when the "e" key is pressed while adjacent to, a textbox will appear with "Hello World" (obviously changeable) on the screen.